% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datoselectorales.R
\docType{data}
\name{alcaldes}
\alias{alcaldes}
\title{"Elecciones municipales en Chile: Alcaldes 2004-2016"}
\format{
Base de datos con 18706 filas y 25 columnas:
\describe{
  \item{\code{tipo_de_eleccion}}{character Tipo de elección. En este caso, de alcaldes.}
  \item{\code{cargo}}{character Cargo al que se postula el candidato/a.}
  \item{\code{fecha_de_eleccion}}{double Fecha exacta en que se llevó a cabo la elección.}
  \item{\code{ano_de_eleccion}}{double Año en que se llevó a cabo la elección.}
  \item{\code{inicio_de_periodo}}{double Inicio del periodo del alcalde o alcaldesa electo.}
  \item{\code{fin_de_periodo}}{double Término del periodo del alcalde o alcaldesa electo.}
  \item{\code{periodo}}{character Periodo en que gobernará el alcalde o alcaldesa electo.}
  \item{\code{id_region}}{double Número identificador de la región.}
  \item{\code{region}}{character Nombre de la región.}
  \item{\code{circunscripcion_senatorial}}{double Circunscripción senatorial.}
  \item{\code{distrito}}{double Distrito electoral.}
  \item{\code{comuna}}{character Nombre de la comuna.}
  \item{\code{circunscripcion_electoral}}{double Circunscripción electoral.}
  \item{\code{nombre_circunscripcion_electoral}}{character Nombre de la circunscripción electoral.}
  \item{\code{sexo_mesa}}{character Sexo de la mesa de votación (hasta las elecciones del 2009).}
  \item{\code{candidato}}{character Nombre completo del candidato/a.}
  \item{\code{nombres}}{character Nombre del candidato/a.}
  \item{\code{apellido_paterno}}{character Apellido paterno del candidato/a.}
  \item{\code{apellido_materno}}{character Apellido materno del candidato/a.}
  \item{\code{sexo}}{character Sexo al que pertenece el candidato/a.}
  \item{\code{electo}}{character Si el candidato o candidata es electo/a o no en las elecciones que compite, o si pasa a segunda vuelta.}
  \item{\code{partido}}{character Partido al que pertenece el candidato/a.}
  \item{\code{sigla_partido}}{character Sigla del partido al que pertenece el candidato/a.}
  \item{\code{lista}}{character Lista en la que va el partido del candidato/a.}
  \item{\code{votos_totales}}{double Votos totales según corresponda.}
}
}
\source{
\url{https://www.servel.cl/resultados-en-excel/}
}
\usage{
alcaldes
}
\description{
"Elecciones municipales en Chile: Alcaldes 2004-2016"
}
\keyword{datasets}
