% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datoselectorales.R
\docType{package}
\name{datoselectorales}
\alias{datoselectorales}
\title{Paquete con datos electorales de Chile}
\description{
Este paquete contiene información eletoral de Chile,
sólo se han fomateado los nombres y algunos partidos
políticos con nombres en distintos formatos.
}
