% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datoselectorales.R
\docType{data}
\name{presidenciales}
\alias{presidenciales}
\title{"Elecciones presidenciales en Chile: 1989-2017"}
\format{
Una base de datos con 64366 filas y 21 columnas:
\describe{
  \item{\code{tipo_de_eleccion}}{character Tipo de elección. En este caso, presidencial.}
  \item{\code{cargo}}{character Cargo al que se postula el candidato/a.}
  \item{\code{fecha_de_eleccion}}{double Fecha exacta en que se llevó a cabo la elección.}
  \item{\code{ano_de_eleccion}}{double Año en que se llevó a cabo la elección.}
  \item{\code{periodo}}{character Periodo en que gobernará el presidente electo.}
  \item{\code{votacion_presidencial}}{character Si los votos corresponden a primera vuelta, segunda vuelta o elección única (solo en las elecciones de 1989 y 1993).}
  \item{\code{id_region}}{double Número idenficiador de la región.}
  \item{\code{region}}{character Nombre de la región.}
  \item{\code{id_provincia}}{double Número idenficiador de la provincia.}
  \item{\code{provincia}}{character Nombre de la provincia.}
  \item{\code{comuna}}{character Nombre de la comuna.}
  \item{\code{sexo_mesa}}{character Sexo de la mesa de votación (hasta las elecciones del 2009).}
  \item{\code{candidato}}{character Nombre completo del candidato o candidata.}
  \item{\code{nombres}}{character Nombres del candidato /a.}
  \item{\code{apellido_paterno}}{character Apellido paterno del candidato/a.}
  \item{\code{apellido_materno}}{character Apellido materno del candidato/a.}
  \item{\code{sexo}}{character Sexo al que pertenece el candidato o candidata.}
  \item{\code{electo}}{character Si el candidato o candidata es electo/a o no en las elecciones que compite, o si pasa a segunda vuelta.}
  \item{\code{partido}}{character Partido al que pertenece el candidato/a.}
  \item{\code{sigla_partido}}{character Sigla del partido al que pertenece el candidato/a.}
  \item{\code{votos_totales}}{double Votos totales según corresponda.}
}
}
\source{
\url{https://www.servel.cl/resultados-en-excel/}
}
\usage{
presidenciales
}
\description{
"Elecciones presidenciales en Chile: 1989-2017"
}
\keyword{datasets}
